Refactor: 클라이언트 프로젝트 구조 수정
