--- src/View/Header/styles.css
@@ -1,30 +1,0 @@
-/* 정렬 참고자료 : 추후 삭제 예정 */
-
-.container {
-    /* 모달창 크기 */
-    width: 300px;
-    height: 200px;
-
-    /* 최상단 위치 */
-    z-index: 999;
-
-    /* 중앙 배치 */
-    /* top, bottom, left, right 는 브라우저 기준으로 작동한다. */
-    /* translate는 본인의 크기 기준으로 작동한다. */
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-
-    /* 모달창 디자인 */
-    background-color: gray;
-    border: 1px solid black;
-    border-radius: 8px;
-}
-
-/* 모달창 내부 X버튼 */
-.close {
-    position: absolute;
-    right: 10px;
-    top: 10px;
-}
