--- src/css/List.css
@@ -1,49 +1,0 @@
-.list.close {
-    animation-duration: 0.2s;
-    animation-name: slide-out;
-    min-width: 30px;
-    width: 30px;
-}
-
-.list.open {
-    animation-duration: 0.2s;
-    animation-name: slide-in;
-    min-width: 100%;
-    width: 100%;
-    height: 100%;
-}
-
-.list .hover-close {
-    align-items: elft;
-    background-color: #aab1bc;
-    border-radius: 50%;
-    cursor: pointer;
-    display: block;
-    height: 32px;
-    justify-content: center;
-    position: absolute;
-    right: -13px;
-    top: 50%;
-    transition: 0.2;
-    width: 32px;
-    z-index: 100;
-}
-
-.hover-close.open.hide {
-    display: none;
-}
-
-.hover-close img {
-    height: 16px;
-    transform: rotate(180deg);
-    width: 16px;
-}
-
-.hover-close.open img {
-    transform: none;
-}
-
-.list .content {
-    height: 100%;
-    width: 100%;
-}
