--- src/custom.d.ts
@@ -1,4 +1,0 @@
-declare module '*.jpg';
-declare module '*.png';
-declare module '*.jpeg';
-declare module '*.gif';
