+++ src/views/webCam/index.css
@@ -1,0 +1,21 @@
+#wrap {
+    position: relative;
+    border: 1px solid brown;
+}
+  
+canvas {
+    position: absolute;
+    top: 50;
+    left: 0;
+    border: 1px solid black;
+}
+
+div#wrap {
+    position: relative;
+}
+
+div#wrap canvas {
+    position: absolute;
+    top: 0;
+    left: 0;
+}
