+++ tsconfig.json
@@ -5,6 +5,7 @@
         "strict": true,
         "esModuleInterop": true,
         "outDir": "dist",
+        "declaration": true,
         "jsx": "react-jsx"
     },
     "include": ["src/**/*"],
